Add support for CADDY_HTTP_PORT in the LMS/CMS configuration for MFEs
In a local testing configuration, people might run "tutor local" with
CADDY_HTTP_PORT set to a non-default value, and HTTPS disabled
(i.e. ENABLE_HTTPS set to false). When that happens, both the LMS/CMS
and the MFEs listen on a port other than the default 80.

This requires that the LMS/CMS configure themselves with MFE URLs that
not only include MFE_HOST, but also refer to CADDY_HTTP_PORT
correctly.

Add that functionality for the LMS/CMS settings patches.